I completed the conversion site with little help!
completed this portion of the lab! when I click convert, it scrolls to the top of the page sometimes. I don't know why. also not sure if IBMs code does that too, as I tried to copy paste their code and it didn't work! Probably outdated. Oh well, maybe I'll come back to it later (: